#master - Authentication and verification done.

+++ src/controller/barber-shop.controller.js
@@ -15,21 +15,30 @@ router.get('/:id', (req, res) => {
 
 
 //Need authentication routes
-// router.get('/:id/rating', security.isAuthenticated(), (req, res) => {
-//     res.send(barberShopRepository.getRating(req.params.id))
-// })
+router.get('/:id/rating', security.isAuthenticated, (req, res) => {
+    res.send(barberShopRepository.getRating(req.params.id))
+})
 
 
 //Administrator routes
-router.post('/', (req, res) => {
+router.post('/', security.isAuthenticated, (req, res) => {
+    if(!req.user.hasRole('ADMIN')) {
+        return res.status(401).send({ error: 'You are not allowed to perform this action' });
+    };
     res.send(barberShopRepository.insert(req.body));
 })
 
-router.put('/:id', (req, res) => {
+router.put('/:id', security.isAuthenticated, (req, res) => {
+    if(!req.user.hasRole('ADMIN')) {
+        return res.status(401).send({ error: 'You are not allowed to perform this action' });
+    };
     res.send(barberShopRepository.update(parseInt(req.params.id), req.body));
 })
 
-router.delete('/:id', (req, res) => {
+router.delete('/:id', security.isAuthenticated, (req, res) => {
+    if(!req.user.hasRole('ADMIN')) {
+        return res.status(401).send({ error: 'You are not allowed to perform this action' });
+    };
     barberShopRepository.remove(parseInt(req.params.id), req.body);
     res.sendStatus(200);
 })
@@ -37,12 +46,17 @@ router.delete('/:id', (req, res) => {
 
 //User routes
 router.post('/:id/schedule', (req, res) => {
+    if(!req.user.hasRole('USER')) {
+        return res.status(401).send({ error: 'You are not allowed to perform this action' });
+    };
     res.send(barberShopRepository.schedule(req.params.id, req.body))
 })
 
 router.post('/:id/rate', (req, res) => {
+    if(!req.user.hasRole('USER')) {
+        return res.status(401).send({ error: 'You are not allowed to perform this action' });
+    };
     res.send(barberShopRepository.rate(req.params.id, req.body))
 })
-
 
 module.exports = router;
